:fire: Removes unused MultiQC parameters
This commit removes commented-out MultiQC-related parameters from the `nextflow.config` file and the corresponding channels from `workflows/tskit.nf`.

The `NfcoreTemplate.email` function call in the `TSKIT` workflow is also updated to remove the `multiqc_report` argument, as it is no longer needed.

These changes aim to clean up the configuration and workflow files by removing parameters and code that are no longer in use, which simplifies maintenance and improves readability.