:bug: allow conditional setting of tsdate --ne argument
This commit enables the conditional setting of the `--ne` argument for the tsdate tool within the TSinfer workflows.

Instead of always passing the `tsdate_ne` parameter to the tsdate command, it is now only included if the `params.tsdate_ne` parameter is defined.

This change provides more flexibility in configuring the tsdate step, allowing users to run the process without needing to define a `tsdate_ne` value.